Implemented RouteListActivity.kt
It is for a dynamic selection of locations for the route option. Then it will transfer you to the ArActivity.kt
The button AR Experience in the Menu is deprecated for now!
