+++ app/src/main/java/activities/MapActivity.kt
@@ -33,7 +33,7 @@ import timber.log.Timber.i
 import java.net.URL
 
 
-class MapActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerDragListener, GoogleMap.OnMarkerClickListener {
+class MapActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerDragListener, GoogleMap.OnMarkerClickListener{
 
     private lateinit var map: GoogleMap
     private lateinit var binding: ActivityMapBinding
@@ -115,107 +115,7 @@ class MapActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerD
         getCurrentLocation()
         setPlacemarks()
 
-
-        val sydney = LatLng(-34.0, 151.0)
-        val opera = LatLng(-33.9320447,151.1597271)
-        map.addMarker(MarkerOptions().position(sydney).title("Marker in Sydney"))
-        map.addMarker(MarkerOptions().position(opera).title("Opera House"))
-
-
-
-// Declare polyline object and set up color and width
-        val options = PolylineOptions()
-        options.color(Color.RED)
-        options.width(5f)
-
-        // build URL to call API
-        val url = getURL(sydney, opera)
-
-        async {
-            // Connect to URL, download content and convert into string asynchronously
-            val result = URL(url).readText()
-            uiThread {
-
-
-                // When API call is done, create parser and convert into JsonObjec
-                val parser = Parser()
-                val stringBuilder: StringBuilder = StringBuilder(result)
-                val json: JsonObject = parser.parse(stringBuilder) as JsonObject
-
-
-                // get to the correct element in JsonObject
-                val routes = json.array<JsonObject>("routes")!!
-                val legs = routes[0]["legs"] as JsonArray<JsonObject>
-                val points = legs [0] ["steps"] as JsonArray<JsonObject>
-
-
-
-                // For every element in the JsonArray, decode the polyline string and pass all points to a List
-                val polypts = points.flatMap { decodePoly(it.obj("polyline")?.string("points")!!)  }
-                i("polypts ${points.flatMap { decodePoly(it.obj("polyline")?.string("points")!!)  }}")
-
-                // Add  points to polyline and bounds
-                options.add(sydney)
-                for (point in polypts) options.add(point)
-                options.add(opera)
-                map.addPolyline(options)
-
-
-            }
-        }
     }
-
-
-
-
-    private fun getURL(from : LatLng, to : LatLng) : String {
-        val origin = "origin=" + from.latitude + "," + from.longitude
-        val dest = "destination=" + to.latitude + "," + to.longitude
-        //val sensor = "sensor=false"
-        val apikey = getString(R.string.google_maps_key)
-        val params = "$origin&$dest&key=$apikey"
-        return "https://maps.googleapis.com/maps/api/directions/json?$params"
-    }
-
-    private fun decodePoly(encoded: String): List<LatLng> {
-        val poly = ArrayList<LatLng>()
-        var index = 0
-        val len = encoded.length
-        var lat = 0
-        var lng = 0
-
-        while (index < len) {
-            var b: Int
-            var shift = 0
-            var result = 0
-            do {
-                b = encoded[index++].code - 63
-                result = result or (b and 0x1f shl shift)
-                shift += 5
-            } while (b >= 0x20)
-            val dlat = if (result and 1 != 0) (result shr 1).inv() else result shr 1
-            lat += dlat
-
-            shift = 0
-            result = 0
-            do {
-                b = encoded[index++].code - 63
-                result = result or (b and 0x1f shl shift)
-                shift += 5
-            } while (b >= 0x20)
-            val dlng = if (result and 1 != 0) (result shr 1).inv() else result shr 1
-            lng += dlng
-
-            val p = LatLng(lat.toDouble() / 1E5,
-                lng.toDouble() / 1E5)
-            poly.add(p)
-        }
-
-        return poly
-    }
-
-
-
 
 
     private fun setPlacemarks() {
@@ -290,6 +190,7 @@ class MapActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerD
     }
 
     override fun onMarkerDrag(marker: Marker) {
+        getAddress(marker)
     }
 
     override fun onMarkerDragEnd(marker: Marker) {
@@ -309,6 +210,8 @@ class MapActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerD
         //location.title = (adress)
 
     }
+
+
 
     override fun onBackPressed() {
 
@@ -359,6 +262,8 @@ class MapActivity : AppCompatActivity(), OnMapReadyCallback, GoogleMap.OnMarkerD
 
     return addressText
     }
+
+
 
 }
